fix: code improvement
