+++ package.json
@@ -11,7 +11,8 @@
   },
   "dependencies": {
     "@expo/webpack-config": "^19.0.0",
-    "@react-native-async-storage/async-storage": "^1.21.0",
+    "@react-native-async-storage/async-storage": "1.18.2",
+    "@react-native-community/datetimepicker": "^7.2.0",
     "@react-navigation/bottom-tabs": "^6.5.11",
     "@react-navigation/native": "^6.1.9",
     "@react-navigation/stack": "^6.3.20",
@@ -25,7 +26,7 @@
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
     "react-native-snap-carousel-v4": "^1.0.1",
-    "react-native-svg": "^14.1.0",
+    "react-native-svg": "13.9.0",
     "react-native-uuid": "^2.0.1",
     "react-native-web": "~0.19.6",
     "react-router-native": "^6.21.3",
